runtime: fix struct Sigaction for Linux/386         We need to use kernel headers to generate defs_linux_$GOARCH.h
R=golang-dev, dave, alex.brainman, iant
CC=golang-dev
https://golang.org/cl/6296091

+++ src/pkg/runtime/defs_linux_386.h
@@ -132,7 +132,7 @@ struct Sigaction {
 	void	*k_sa_handler;
 	uint32	sa_flags;
 	void	*sa_restorer;
-	uint32	sa_mask;
+	uint64	sa_mask;
 };
 struct Siginfo {
 	int32	si_signo;
